Move `Messages` and related types to new module
